Replace custom OAuth auth module with built-in OIDC provider
  Thunder IDP now supports standard OIDC with well-known discovery endpoints,
  refresh tokens, and standard claims. This replaces the custom OAuth
  implementation with Backstage's built-in oidcAuthenticator.

  Changes:
  - Rename plugin: auth-backend-module-openchoreo-default → openchoreo-idp
  - Simplify auth module from ~380 lines to ~90 lines using oidcAuthenticator
  - Remove custom JWT utilities (jwtUtils.ts) - OIDC handles token validation
  - Remove pseudo-refresh token workaround - Thunder now returns real refresh tokens
  - Rename provider ID: default-idp → openchoreo-idp for clarity
  - Update config to use metadataUrl (OIDC discovery) instead of manual URLs
  - Add session.secret requirement for OIDC cookie-based sessions

  Breaking changes:
  - Provider ID changed from 'default-idp' to 'openchoreo-idp'
  - Config structure changed: authorizationUrl/tokenUrl → metadataUrl
  - Thunder callback URL must be updated to /api/auth/openchoreo-idp/handler/frame
  - SESSION_SECRET env var now required

Fixes: https://github.com/openchoreo/openchoreo/issues/1198
