+++ app-config.yaml
@@ -86,19 +86,19 @@ techdocs:
 auth:
   # see https://backstage.io/docs/auth/ to learn about auth providers
   environment: development
+  # session.secret is required for OIDC - set via SESSION_SECRET env var or in app-config.local.yaml
   providers:
-    # OpenChoreo Default IDP - Custom OAuth provider (without discovery endpoint)
+    # OpenChoreo IDP - OIDC provider with discovery endpoint
     # Environment variables are injected by Helm chart (see https://github.com/openchoreo/openchoreo install/helm/openchoreo/templates/backstage/deployment.yaml)
     # Client credentials are automatically configured from asgardeoThunder.backstagePortal values
     # For local development with Kind: URLs set via https://github.com/openchoreo/openchoreo install/dev/openchoreo-values.yaml (e.g., http://sts.openchoreo.localhost)
     # For production: set via https://github.com/openchoreo/openchoreo install/helm/openchoreo/values.yaml backstage.thunder.baseUrl (external URL with ingress/load balancer)
-    default-idp:
+    openchoreo-idp:
       development:
         clientId: ${OPENCHOREO_AUTH_CLIENT_ID}
         clientSecret: ${OPENCHOREO_AUTH_CLIENT_SECRET}
-        authorizationUrl: ${OPENCHOREO_AUTH_AUTHORIZATION_URL}
-        tokenUrl: ${OPENCHOREO_AUTH_TOKEN_URL}
-        scope: 'openid profile email'
+        metadataUrl: ${THUNDER_BASE_URL}/.well-known/openid-configuration
+        prompt: auto
 
     # Guest provider - used when openchoreo.features.auth.enabled is false
     # Allows users to access the portal without authentication (demo/development mode)
