+++ packages/backend/package.json
@@ -41,7 +41,7 @@
     "@backstage/plugin-search-backend-node": "1.3.15",
     "@backstage/plugin-techdocs-backend": "2.1.0",
     "@backstage/plugin-user-settings-backend": "0.3.7",
-    "@openchoreo/backstage-plugin-auth-backend-module-openchoreo-default": "workspace:^",
+    "@openchoreo/backstage-plugin-auth-backend-module-openchoreo-idp": "workspace:^",
     "@openchoreo/backstage-plugin-backend": "workspace:^",
     "@openchoreo/backstage-plugin-catalog-backend-module": "workspace:^",
     "@openchoreo/backstage-plugin-catalog-backend-module-openchoreo-users": "workspace:^",
