--- plugins/auth-backend-module-openchoreo-default/src/index.ts
@@ -1,8 +1,0 @@
-/***/
-/**
- * The openchoreo-default backend module for the auth plugin.
- *
- * @packageDocumentation
- */
-
-export { OpenChoreoDefaultAuthModule } from './auth';
