+++ plugins/auth-backend-module-openchoreo-idp/package.json
@@ -1,9 +1,9 @@
 {
-  "name": "@openchoreo/backstage-plugin-auth-backend-module-openchoreo-default",
+  "name": "@openchoreo/backstage-plugin-auth-backend-module-openchoreo-idp",
   "version": "0.1.0",
   "license": "Apache-2.0",
   "private": true,
-  "description": "The openchoreo-default backend module for the auth plugin.",
+  "description": "OpenChoreo IDP authentication module for Backstage using OIDC.",
   "main": "src/index.ts",
   "types": "src/index.ts",
   "publishConfig": {
@@ -13,7 +13,8 @@
   },
   "backstage": {
     "role": "backend-plugin-module",
-    "pluginId": "auth"
+    "pluginId": "auth",
+    "pluginPackage": "@backstage/plugin-auth-backend"
   },
   "scripts": {
     "start": "backstage-cli package start",
@@ -28,14 +29,11 @@
     "@backstage/backend-plugin-api": "1.4.3",
     "@backstage/catalog-model": "1.7.5",
     "@backstage/plugin-auth-backend-module-oidc-provider": "0.4.8",
-    "@backstage/plugin-auth-node": "0.6.8",
-    "jose": "^6.1.3",
-    "passport-oauth2": "1.8.0"
+    "@backstage/plugin-auth-node": "0.6.8"
   },
   "devDependencies": {
     "@backstage/backend-test-utils": "1.9.0",
-    "@backstage/cli": "0.34.3",
-    "@types/passport-oauth2": "1.8.0"
+    "@backstage/cli": "0.34.3"
   },
   "files": [
     "dist"
