+++ plugins/auth-backend-module-openchoreo-idp/src/index.ts
@@ -1,0 +1,7 @@
+/**
+ * OpenChoreo IDP authentication module for Backstage.
+ *
+ * @packageDocumentation
+ */
+
+export { openChoreoIdpAuthModule } from './module';
